+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **?.?.? Changes:**
 
 * Twitch Voting:
+  * Fixed voting not being initialized properly if the login command is used during a run
   * Username and OAuth token are no longer reset on mod update, you should now only have to log in once, ever.
 
 <details>